Feature: User configuration for behavior on clicking repo title (#52)
### Summary

In this PR, a new user configuration setting is available on the
settings page. The new setting allows the user to pick the behavior they
wish to happen when clicking the empty space in between the name of the
repo and the dropdown icon for each repository section.

Previously it would just link you to the repo, but some people preferred
to be able to expand the section instead. This is the compromise that I
came up with that lets the user choose. If there is an error in fetching
the settings, it will default to expanding.

### Changes
- New settings section to set the header click behavior value
- New extension functions to handle getting and setting this header
click behavior value
- New hook in React created to fetch this value on the popup