Feature: Review State on Pull Request Display (#57)
### Summary

In this PR, the user will now be able to see what pull requests they
have recently reviewed. The latest review given by the user is fetched
from the GitHub API on each pull request, and based on the type of
review they gave, a different icon will be presented.

The currently supported review states that will show icons are for if
the user left comments, requested changes, or approved it. It is
important to note that the icons don't exactly match up with the icon's
in GitHub but are close. To give some clarity, a tooltip will display
explaining what the icon means.

### Changes
- Added a new Icons to the pull request component in the PRDisplay page