Feature: Status Check Setting (#61)
### Summary

In this PR, a new setting has been introduced for the status check
feature. By default this setting will be turned off and there is a short
description that it may not be completely accurate. More details on why
it is labeled as an "experimental" feature is described in #59

### Changes
- Added a new hook along with extension methods to save and fetch the
configured setting from storage
- Hook is used in the <PRDisplay /> component and passed to each of the
<PullRequest /> components
- Updates in <PullRequest />  components to make use of the new setting